refactor: allow rolling over to start again
The introduction of the SubstanceSituation class has now allowed us to do away
with the verse class completely. Because the SubstanceSituation can tell us
which situation comes next, we can use that to progress the song. This allows
a new feature that was always suggested by the lyrics but never catered for:
starting again from 99 when we reach the end of the song.

In addition, the SubstanceSong class (now renamed to remove the last hard-coded)
mention of beer) is instantiated, to underscore the fact that this is "a"
song, not "the" song; we could now quite easily create new songs about
new substances, where the actions taken create different new situations.